feat(main) refOrder의 트랜젝션을 위한 api만드는 중, sql문 동작을 post맨으로   확인 한 뒤 view랑 연결하면 됨, 그 뒤 ppt 만들고 지원하고 로그인 회원가입은 내일 만들자

+++ just_board_api/src/main/java/com/just_board/just_board/entity/Comment.java
@@ -37,11 +37,11 @@ public class Comment {
     // private Long articleId;
     private Date date;
     @Column(name="ref_id")
-    private int refId;
-    private int step;
-    private int ref;
+    private Long refId;
+    private Long step;
+    private Long ref;
     @Column(name="ref_order")
-    private int refOrder;
+    private Long refOrder;
     @Column(name="answer_num")
     private Long answerNum;    
 }
